Feat: Table A81 Related
A81

+++ src/Core/Application/ReportGenerator/Teachers/ATeacherDto.cs
@@ -5,7 +5,7 @@ public class ATeacherDto
     public string Teacher { get; set; }
     public string? TeacherEnglishName { get; set; }
     public string? Degree { get; set; }
-    public decimal? DegreeYear { get; set; }
+    public short? DegreeYear { get; set; }
     public string? Responsibilities { get; set; }
     public string? Qualification { get; set; }
     public string? WorkType { get; set; }
